后台管理-完成评论部分更新操作、递归删除

+++ blog-cms/src/views/comment/CommentList.vue
@@ -40,20 +40,18 @@
 				</el-table-column>
 				<el-table-column label="是否公开" width="80">
 					<template v-slot="scope">
-						<el-switch v-model="scope.row.published" @change=""></el-switch>
+						<el-switch v-model="scope.row.published" @change="commentPublishedChanged(scope.row)"></el-switch>
 					</template>
 				</el-table-column>
 				<el-table-column label="邮件提醒" width="80">
 					<template v-slot="scope">
-						<el-switch v-model="scope.row.notice" @change=""></el-switch>
+						<el-switch v-model="scope.row.notice" @change="commentNoticeChanged(scope.row)"></el-switch>
 					</template>
 				</el-table-column>
-				<el-table-column label="操作" width="190">
+				<el-table-column label="操作" width="200">
 					<template v-slot="scope">
 						<el-button type="primary" icon="el-icon-edit" size="mini" @click="">编辑</el-button>
-						<el-popconfirm title="确定删除吗？" icon="el-icon-delete" iconColor="red" @onConfirm="deleteCommentById(scope.row.id)">
-							<el-button size="mini" type="danger" icon="el-icon-delete" slot="reference">删除</el-button>
-						</el-popconfirm>
+						<el-button type="danger" icon="el-icon-delete" size="mini" @click="deleteCommentById(scope.row.id)">删除</el-button>
 					</template>
 				</el-table-column>
 			</el-table>
@@ -69,7 +67,7 @@
 
 <script>
 	import Breadcrumb from "@/components/Breadcrumb";
-	import {getCommentListByQuery, getBlogList} from '@/network/comment'
+	import {getCommentListByQuery, getBlogList, updatePublished, updateNotice, deleteCommentById} from '@/network/comment'
 
 	export default {
 		name: "CommentList",
@@ -158,25 +156,48 @@
 					this.msgError("请求失败")
 				})
 			},
-			deleteCommentById(id) {
-				deleteCommentById(id).then(res => {
-					console.log(res)
+			//切换评论邮件提醒状态
+			commentNoticeChanged(row) {
+				updateNotice(row.id, row.notice)
+				.then(res => {
 					if (res.code === 200) {
-						this.msgSuccess(res.msg)
-						this.getCommentList()
+						this.msgSuccess(res.msg);
 					} else {
 						this.msgError(res.msg)
 					}
 				}).catch(() => {
 					this.msgError("请求失败")
 				})
+			},
+			deleteCommentById(id) {
+				this.$confirm('此操作将永久删除该评论<strong style="color: red">及其所有子评论</strong>，是否删除?', '提示', {
+					confirmButtonText: '确定',
+					cancelButtonText: '取消',
+					type: 'warning',
+					dangerouslyUseHTMLString: true
+				}).then(() => {
+					deleteCommentById(id).then(res => {
+						console.log(res)
+						if (res.code === 200) {
+							this.msgSuccess(res.msg)
+							this.getCommentList()
+						} else {
+							this.msgError(res.msg)
+						}
+					}).catch(() => {
+						this.msgError("请求失败")
+					})
+				}).catch(() => {
+					this.$message({
+						type: 'info',
+						message: '已取消删除'
+					});
+				});
 			}
 		}
 	}
 </script>
 
 <style scoped>
-	.el-button {
-		margin-right: 10px;
-	}
+
 </style>
